+++ src/renderer/App.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { onMounted } from 'vue';
-import AppMenu from './components/AppMenu.vue';
-import AppHeader from './components/AppHeader.vue';
+import AppMenu from './components/Menu.vue';
+import AppHeader from './components/Header.vue';
 
 onMounted(async () => {
   const isDark = await window.electronAPI.theme.isDark();
